Bug detected wen plotting data sets with more than 12 variables using plot3D function in the GUI.
Sol.: the invocation to the plot3D function has been modified in all the instances in Alignment. Namely, we indicate the number of rows to vsualize in the plot.

In addition, the synchronization structure that can be saved into a file or to the workspace has been simplified. Every time that a new synchronization method is selected, the internal structure of MVBatch is initialized to avoid having parameters from previous synchronizations.